src/requirements.txt
openai==1.54.1
python-dotenv==1.0.1
requests==2.32.3
resend==2.4.0
stripe==6.6.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-anthropic==0.2.0
langchain-core==0.3.0
langchain-mcp-tools==0.2.0
pydantic==2.7.1
uvx==0.1.11
